requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-filter>=24.0
drf-spectacular>=0.27.0
drf-spectacular-sidecar>=2024.1.1
scikit-learn>=1.4.0
numpy>=1.26.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
argon2-cffi>=23.1.0
gunicorn>=21.0.0
dj-database-url>=2.1.0
whitenoise>=6.6.0
